requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1

requests==2.31.0
httpx==0.27.0
aiofiles==23.2.1

llama-index==0.10.30
llama-index-embeddings-huggingface==0.1.4
llama-index-llms-groq==0.1.3
llama-index-vector-stores-pinecone==0.1.6

pinecone-client==3.2.2

sentence-transformers==2.6.1
torch==2.2.2

pypdf==4.2.0
python-docx==1.1.0

gunicorn==21.2.0